+++ Sources/Paywall/Paywall/Paywall View Controller/SWPaywallViewController.swift
@@ -246,11 +246,16 @@ final class SWPaywallViewController: UIViewController, SWWebViewDelegate {
 
 	override func viewWillDisappear(_ animated: Bool) {
 		super.viewWillDisappear(animated)
-		if isPresented,
-      !isSafariVCPresented,
-      !calledDismiss {
-				Paywall.delegate?.willDismissPaywall?()
-		}
+    guard isPresented else {
+      return
+    }
+    if isSafariVCPresented {
+      return
+    }
+    if calledDismiss {
+      return
+    }
+    Paywall.delegate?.willDismissPaywall?()
 	}
 
 	override func viewDidDisappear(_ animated: Bool) {